Re-export UEH service
